// src/frontend/src/backend-shim.d.ts
export interface UserProfile {
  subjects: Array<string>;
  educationClass?: string;
  name: string;
  level: string;
}